% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Titlenof the data}
\format{
A data frme with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollar}
\item{carat}{weight of diamonds, in carat}
...
}
}
\source{
where you get the data
}
\usage{
key_crop_yields
}
\description{
More details about the data
}
\section{Background}{

The data was collected from ... over the course of ...
}

\examples{
print(key_crop_yields)
}
\keyword{datasets}
