% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_normalise.R
\name{stack_normalize}
\alias{stack_normalize}
\alias{stack_normalise}
\title{Convert numbers into proportions}
\usage{
stack_normalize(x)

stack_normalise(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
A vector od propertions
}
\description{
Title
}
\examples{
x <- c(10, 30, 40)
stack_normalize(x)
x <- c(75, 0, 5, 20, NA)
stack_normalize(x)
stack_normalise(x)
}
